WIP: flow operations through host, separate utils
